requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
pytest 
pytest-django 
pytest-cov